django-ninja/[[ cookiecutter.project_slug ]]/requirements.txt
Django==5.2.2
django-ninja==1.4.3
django-ninja-extra==0.30.1
ninja-schema==0.14.2
django-oauth-toolkit==3.0.1
django-jazzmin==3.0.1
gunicorn==23.0.0
l4py==0.1.9
royman-dotenv==1.1.2
whitenoise==6.9.0